Sets focus when starting screen
On NewGameController, sets the focus back to the Game Name TextField.
This way when you navigate away from the screen and back, or start it
for the first time, the focus will always be on the same thing.

On LoadGameController, set the focus on Back button since its the
default when going to the screen. Now when you navigate away and back,
it'll still be on back.